requirements.txt
blinker==1.9.0
click==8.3.1
customtkinter==5.2.2
darkdetect==0.8.0
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
pillow==12.1.1
PyMySQL==1.1.2
Werkzeug==3.1.5
